resolves circular dependency in lib
