+++ packages/lib/src/consoleOnce.ts
@@ -1,0 +1,7 @@
+import once from 'lodash/once';
+
+export const consoleOnce = {
+  error: once(console.error),
+  warn: once(console.warn),
+  log: once(console.log),
+};
